\name{billy_inp_p}
\alias{billy_inp_p}
\docType{data}
\title{
Billy's Bus Company Data: Inputs Prices
}
\description{
Data set used for example 4.7 and 4.9.3, p.114 and p.124, of \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. 
}
\usage{data("multil_data")}
\format{
  A data frame with 5 observations on the following 4 variables.
  \describe{
    \item{\code{year}}{years of observation}
    \item{\code{p1}}{a numeric vector containing prices of labor}
    \item{\code{p2}}{a numeric vector containing prices of capital}
    \item{\code{p3}}{a numeric vector containing prices of other inputs}
  }
}
\source{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.
}
\references{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.
}
\keyword{datasets}
