\name{dat.p}
\alias{dat.p}
\docType{data}
\title{
Transitive Indexes in Panel Data: Simulated Regional Data on Labor Inputs Prices
}
\description{
A simulated data set in panel data format that can be used to create mst or eks indexes. It contains information on hourly wages for three categories of workers.
}
\usage{data("multil_data")}
\format{
  A data frame with 80 observations on the following 5 variables.
  \describe{
    \item{\code{year}}{temporal dimension of the panel}
    \item{\code{region}}{id of the region}
    \item{\code{x1}}{hourly wage of of labor input 1}
    \item{\code{x2}}{hourly wage of of labor input 2}
    \item{\code{x3}}{hourly wage of of labor input 3}
  }
}
\keyword{datasets}
