\name{dat.q}
\alias{dat.q}
\docType{data}
\title{
Transitive Indexes in Panel Data: Simulated Regional Data on Labor Inputs Quantities
}
\description{
A simulated data set in panel data format that can be used to create mst or eks indexes. It contains hours wokred for threee categories of workers.
}
\usage{data("multil_data")}
\format{
  A data frame with 80 observations on the following 5 variables.
  \describe{
    \item{\code{year}}{temporal dimension of the panel}
    \item{\code{region}}{id of the region}
    \item{\code{x1}}{hours worekd by labor input 1}
    \item{\code{x2}}{hours worekd by labor input 2}
    \item{\code{x3}}{hours worekd by labor input 3}
  }
}
\keyword{datasets}
