\name{I_nt}
\alias{I_nt}
\docType{data}
\title{
Matrix of bilateral Fisher fixed-base comparisons
}
\description{
Matrix of bilateral quantity Fisher comparisons for region 1 and region 2 and for 2014 and 2015 of the Simulated Regional Data on Labor Inputs Statistics
}
\usage{data("multil_data")}
\format{
  The format is:
 num [1:4, 1:4] 1 0.896 0.179 0.137 1.116 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:4] "1.2014" "1.2015" "2.2014" "2.2015"
  ..$ : chr [1:4] "1.2014" "1.2015" "2.2014" "2.2015"
}

\source{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.}
\references{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.}

\keyword{datasets}
