\name{rail_out_p}
\alias{rail_out_p}
\docType{data}
\title{
Australian National Railways Data: Output Prices
}
\description{
Data set used for example 4.10, p.128, of \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition.  
}
\usage{data("multil_data")}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{p1}}{prices of Mainland Freight services ($/NTK)}
    \item{\code{p2}}{prices of Tasrail Freight services ($/NTK)}
    \item{\code{p3}}{prices of Passenger services ($/PTK) }
  }
}

\source{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.}
\references{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.}

\keyword{datasets}
