\name{rail_out_q}
\alias{rail_out_q}
\docType{data}
\title{
Australian National Railways Data: Output Quantities
}
\description{
Data set used for example 4.10, p.128, of \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition.  
}
\usage{data("multil_data")}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{x1}}{quantities of Mainland Freight services (1,000 NTKs)}
    \item{\code{x2}}{quantities of Tasrail Freight services (1,000 NTKs)}
    \item{\code{x3}}{quantities of Passenger services (1,000 PTKs)}
  }
}

\source{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.}
\references{
T. J. Coelli, D. S. Prasada Rao, C. J. O'Donnell, and G. E. Battese. \emph{An introduction to Efficiency and Productivity Analysis}, 2nd edition. Springer Science + Business Media, New York, 2005.}

\keyword{datasets}
